Precompute KPIs for faster loading of results page
